reorganized tool-existing related tests
